cosmetic changes & compiler warnings fixes

+++ EDRSandblast/Drivers/DriverGDRV.c
@@ -19,7 +19,7 @@
 #endif
 
 /*
-* "gdrv.sys" (SHA256: xxx)
+* "gdrv.sys" (SHA256: 31f4cfb4c71da44120752721103a16512444c13c2ac2d857a7e6f13cb679b427)
 */
 
 struct GDRV_MEMORY_READ {
@@ -75,8 +75,8 @@ VOID ReadMemoryPrimitive_GDRV(SIZE_T Size, DWORD64 Address, PVOID Buffer) {
     //copy Size bytes from Src to Dest
     struct GDRV_MEMORY_READ ReadCommand = { 0 };
     ReadCommand.Src = Address;
-    ReadCommand.Dst = Buffer;
-    ReadCommand.ReadSize = Size;
+    ReadCommand.Dst = (DWORD64) Buffer;
+    ReadCommand.ReadSize = (DWORD) Size;
 
     DWORD BytesReturned=0;
     DeviceIoControl(GetDriverHandle_GDRV(),
@@ -104,9 +104,9 @@ VOID WriteMemoryPrimitive_GDRV(SIZE_T Size, DWORD64 Address, PVOID Buffer) {
     }
     //copy Size bytes from Dest to Src
     struct GDRV_MEMORY_WRITE WriteCommand = { 0 };
-    WriteCommand.Src = Buffer;
+    WriteCommand.Src = (DWORD64) Buffer;
     WriteCommand.Dst = Address;
-    WriteCommand.WriteSize = Size;
+    WriteCommand.WriteSize = (DWORD) Size;
 
     DWORD BytesReturned = 0;
     DeviceIoControl(GetDriverHandle_GDRV(),
